+++ FRCore/FRCore/Keychain/KeychainService.swift
@@ -2,7 +2,7 @@
 //  KeychainService.swift
 //  FRCore
 //
-//  Copyright (c) 2020 ForgeRock. All rights reserved.
+//  Copyright (c) 2020 - 2021 ForgeRock. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -17,7 +17,7 @@ public struct KeychainService {
     // MARK: - Property
     
     /// Options for Keychain Service operation
-    fileprivate var options: KeychainOptions
+    public var options: KeychainOptions
     /// SecuredKey used for encryption/decryption of data in Keychain Service
     fileprivate var securedKey: SecuredKey?